Correção de Validação dos Status e inclusão do log4j2.xml

+++ BibliotecaVirtualRestAPI/src/main/java/br/com/biblioteca/rest/api/services/StatusService.java
@@ -3,27 +3,37 @@ package br.com.biblioteca.rest.api.services;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.biblioteca.rest.api.controllers.StatusController;
 import br.com.biblioteca.rest.api.dtos.StatusDTO;
 import br.com.biblioteca.rest.api.enuns.ValidacaoEnum;
+import br.com.biblioteca.rest.api.models.Publisher;
 import br.com.biblioteca.rest.api.models.Status;
+import br.com.biblioteca.rest.api.repositories.PublisherRepository;
 import br.com.biblioteca.rest.api.repositories.StatusRepository;
 
 @Service
 public class StatusService {
 
+	private static final Logger logger = LogManager.getLogger(StatusController.class);
+
 	private StatusRepository repository;
+	private PublisherRepository repositoryPublisher;
 
 	@Autowired
-	public StatusService(StatusRepository repository) {
+	public StatusService(StatusRepository repository, PublisherRepository repositoryPublisher) {
 		super();
 		this.repository = repository;
+		this.repositoryPublisher = repositoryPublisher;
 	}
 
 	public StatusDTO salvar(Status status) {
-		String validacao = validarStatus(status);
+		logger.info("Salvando Status " + status.getDescription() + ".");
+		String validacao = validarStatus(status, true);
 		if (validacao.equals(ValidacaoEnum.OK.getDescricao())) {
 			status.setNemotechnic(status.getNemotechnic().toUpperCase());
 			status.setIsActive(ValidacaoEnum.SIM.getDescricao());
@@ -34,6 +44,7 @@ public class StatusService {
 	}
 
 	public List<StatusDTO> listarTodos() {
+		logger.info("Listando todos os Status.");
 		List<Status> todosStatus = repository.findAll();
 		List<StatusDTO> dtos = new ArrayList<StatusDTO>();
 		for (Status status : todosStatus) {
@@ -43,6 +54,7 @@ public class StatusService {
 	}
 
 	public StatusDTO listarPorId(long id) {
+		logger.info("Listando Status pelo ID.");
 		Status status = repository.findById(id);
 		if (status == null) {
 			return montarDTOErro(null, ValidacaoEnum.NAO_EXISTE.getDescricao());
@@ -56,13 +68,14 @@ public class StatusService {
 		if (statusEditar == null) {
 			return montarDTOErro(null, ValidacaoEnum.NAO_EXISTE.getDescricao());
 		} else {
-			if (validarStatus(status) == ValidacaoEnum.OK.getDescricao()) {
+			if (validarStatus(status, false) == ValidacaoEnum.OK.getDescricao()) {
+				logger.info("Editando Status ." + status.getDescription() + ".");
 				statusEditar.setDescription(status.getDescription());
 				statusEditar.setNemotechnic(status.getNemotechnic().toUpperCase());
 				statusEditar.setIsActive(status.getIsActive());
 				return montarDTOSucesso(repository.save(statusEditar), true, ValidacaoEnum.EDITAR.getDescricao());
 			} else {
-				return montarDTOErro(null, validarStatus(status));
+				return montarDTOErro(null, validarStatus(status, false));
 			}
 		}
 	}
@@ -72,12 +85,17 @@ public class StatusService {
 		if (status == null) {
 			return montarDTOErro(null, ValidacaoEnum.NAO_EXISTE.getDescricao());
 		} else {
-			repository.deleteById(id);
-			return montarDTOSucesso(null, true, ValidacaoEnum.DELETAR.getDescricao());
+			if (validarDeletar(status)) {
+				logger.info("Deletando Status " + status.getDescription() + ".");
+				repository.deleteById(id);
+				return montarDTOSucesso(null, true, ValidacaoEnum.DELETAR.getDescricao());
+			}
+			return montarDTOErro(null, ValidacaoEnum.FK.getDescricao());
 		}
 	}
 
 	public StatusDTO montarDTOSucesso(Status status, Boolean preencheMsgSucesso, String validacao) {
+		logger.info("Montando objeto de sucesso.");
 		StatusDTO dto = new StatusDTO();
 		if (status != null) {
 			dto.setId(status.getId());
@@ -97,6 +115,7 @@ public class StatusService {
 	}
 
 	public StatusDTO montarDTOErro(Status status, String validacao) {
+		logger.info("Montando objeto de erro.");
 		StatusDTO dto = new StatusDTO();
 		if (validacao.equals(ValidacaoEnum.EXISTE.getDescricao())) {
 			dto.setMsgError("Já existe um Status com o nemotécnico " + status.getNemotechnic() + ".");
@@ -110,10 +129,14 @@ public class StatusService {
 		if (validacao.equals(ValidacaoEnum.NAO_EXISTE.getDescricao())) {
 			dto.setMsgError("O Status não existe.");
 		}
+		if (validacao.equals(ValidacaoEnum.FK.getDescricao())) {
+			dto.setMsgError("O Status não pode ser deletado por estar associado a outro retistro.");
+		}
 		return dto;
 	}
 
 	public StatusDTO validarStatusExistente(Status status) {
+		logger.info("Validando se o Status " + status.getDescription() + " já existe.");
 		Status statusValidado = repository.findByNemotechnic(status.getNemotechnic());
 		if (statusValidado != null) {
 			return montarDTOErro(status, ValidacaoEnum.EXISTE.getDescricao());
@@ -122,15 +145,35 @@ public class StatusService {
 		}
 	}
 
-	public String validarStatus(Status status) {
+	public String validarStatus(Status status, boolean salvar) {
+		logger.info("Validando campos do Status.");
 		try {
-			if (status.getDescription().trim().isEmpty() || status.getNemotechnic().trim().isEmpty()) {
-				return ValidacaoEnum.BRANCO.getDescricao();
+			if (salvar) {
+				if (status.getDescription().trim().isEmpty() || status.getNemotechnic().trim().isEmpty()) {
+					return ValidacaoEnum.BRANCO.getDescricao();
+				} else {
+					return ValidacaoEnum.OK.getDescricao();
+				}
 			} else {
-				return ValidacaoEnum.OK.getDescricao();
+				if (status.getDescription().trim().isEmpty() || status.getNemotechnic().trim().isEmpty()
+						|| status.getIsActive().isEmpty()) {
+					return ValidacaoEnum.BRANCO.getDescricao();
+				} else {
+					return ValidacaoEnum.OK.getDescricao();
+				}
 			}
 		} catch (NullPointerException erro) {
 			return ValidacaoEnum.NULL.getDescricao();
+		}
+	}
+
+	public boolean validarDeletar(Status status) {
+		logger.info("Validando se o Status " + status.getDescription() + " está sendo usado.");
+		List<Publisher> publishers = repositoryPublisher.findByStatus(status);
+		if (publishers.size() != 0) {
+			return false;
+		} else {
+			return true;
 		}
 	}
